change iter interface
now takes two functions, `lenfn` which takes a `buf` and returns the
pair `hlen` and `plen`; and `pfn` which takes a `buf` and returns
an `'a`, ie., well-types it. this leaves it to the caller to decide
if to split into hdr,body; and also enables formats where the header
length cannot be known until the header is partially parsed.