Do not depend explicitly on deprecated ppx_driver
To avoid having to pull in a deprecated package depend on either
`ppx_driver` or `ppxlib`, apparently the only reason we depend on it is
to ensure the version is new enough, ppx_cstruct does not require it
directly.

I kept both ppx_driver and ppxlib so that the dependency works with
older compiler versions too.

Do not put version constraints on `ppxlib`, any version is newer than
`ppx_driver`, and if the version constraint is not satisfied then opam
would fall back installing a new `ppx_driver` instead of upgrading an
existing `ppxlib`.

Signed-off-by: Edwin Török <edwin@etorok.net>