Sexplib is now an optional library for the base `Cstruct` module.
A new `Cstruct_sexp` module has been introduced with the serialiser
functions, contained within the `cstruct-sexp` opam package.

To convert old code, simply use `Cstruct_sexp.t` instead of
`Cstruct.t` in a record type for which you are using `[@@deriving sexp]`.
This is a type alias to `Cstruct.t` but also has the right
sexp-conversion functions in scope.  There is an example of this
in the `ppx_test/with-sexp` directory in the source repo.

fixes #222