ppx_tests: preprocess output to remove file id
We primarily just want to check in the ppx unit tests that the
error message is correct. The line number information varies
by compiler version, so we just preprocess to remove that.

In the longer term, this can be fixed more systematically by
using the toplevel_expect_tests package (suggested by @jeremiedimino)
but this PR unblocks the release.